feat: add tsconfig.json
